+++ packages/administration/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nuxt-customer-portal/administration
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [0f05ccf]
+  - @nuxt-customer-portal/ui@0.4.2
+  - @nuxt-customer-portal/core@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes