+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @nuxt-customer-portal/kit
 
+## 0.4.2
+
 ## 0.4.1
 
 ### Patch Changes